requirements/base.txt
# Base requirements for installation *except* Django
#
# Isolating Django from the other requirements allows better cross-version
# (Django) testing.

mysqlclient==1.4.6                      # MySQL client, version 2 will drop Python 2 support
pysolr>=3.3.0,<3.4.0

pytz>=2014,<2015
argparse==1.2.1
python-dateutil==2.4.1
six==1.14.0
polib==1.0.8                            # Translation file loading

Pillow>=3.0.0
sorl-thumbnail>=12.4,<13.0
textract==1.6.3
Unidecode==1.0.22
parsedatetime==2.2

django-tastypie==0.14.3
django-tablib>=0.9.11,<4.0
django-crispy-forms==1.7.2
django-tinymce>=2.6.0,<2.7.0
django-wysiwyg>=0.7.0,<1.0
django-haystack==2.7.0
django-autoslugged==2.0.0.dev0
django-modeltranslation==0.14.3
django-fsm==2.7.0


django-extensions==2.2.9
defusedxml==0.5.0

dicttoxml==1.7.4

#git+https://github.com/mPowering/orb-api-python.git@v1.1.0#orb_api
git+https://github.com/wellfire/orb-api-python.git@python3#orb_api

Markdown==2.6.9
enum34==1.1.6
typing==3.6.1

pathlib2==2.3.3

django-ckeditor==5.9.0

django-webpack-loader==0.6.0                    # webpack loader for webpack-driven application front-end
